Some Name Refactoring.
I changed the file names so that they sort a little better in my nerd tree, so dumb to do it this way but I don't want to monkey with paths in my folders right now.  who knows maybe I will get really ambitious in the next 5 minutes and do it right, but then again maybe not.